feat: rename prismAgent references to edgeAgent
