+++ edge-agent-sdk/src/commonTest/kotlin/org/hyperledger/identus/walletsdk/edgeagent/protocols/issueCredential/OfferCredentialTest.kt
